requirements.txt
# Data Engineering
pyspark>=3.4.0
delta-spark>=2.4.0
databricks-sdk>=0.12.0

# DBT
dbt-core>=1.7.0
dbt-databricks>=1.7.0

# Data ingestion
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
lxml>=5.0.0
html5lib>=1.1
beautifulsoup4>=4.12.0

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
